# Support system environment variables