--- a/lib/importproject.cpp
+++ b/lib/importproject.cpp
@@ -96,7 +96,7 @@ void ImportProject::FileSettings::setDefines(std::string defs)
     defines.swap(defs);
 }
 
-static bool simplifyPathWithVariables(std::string &s, const std::map<std::string, std::string> &variables)
+static bool simplifyPathWithVariables(std::string &s, std::map<std::string, std::string> &variables)
 {
     std::set<std::string> expanded;
     std::string::size_type start = 0;
@@ -108,10 +108,28 @@ static bool simplifyPathWithVariables(std::string &s, const std::map<std::string
         if (expanded.find(var) != expanded.end())
             break;
         expanded.insert(var);
-        std::map<std::string, std::string>::const_iterator it1 = variables.find(var);
-        if (it1 == variables.end())
-            break;
-        s = s.substr(0,start) + it1->second + s.substr(end+1);
+		std::map<std::string, std::string>::const_iterator it1 = variables.find(var);
+		std::string variableValue;
+		// variable was not found within defined variables
+		if (it1 != variables.end()) {
+			variableValue = it1->second;
+		}
+		// also search environment variables
+		else {
+			char const* envValue = std::getenv(var.c_str());
+			if (envValue) {
+				variableValue = std::string(envValue);
+				// also cache value for next time to reduce system access
+				variables[var] = variableValue;
+			}
+			else {
+				// remember that there is no value other than the variable name
+				variables[var] = var;
+				break;
+			}
+		}
+
+		s = s.substr(0, start) + variableValue + s.substr(end + 1);
     }
     if (s.find("$(") != std::string::npos)
         return false;
@@ -119,7 +137,7 @@ static bool simplifyPathWithVariables(std::string &s, const std::map<std::string
     return true;
 }
 
-void ImportProject::FileSettings::setIncludePaths(const std::string &basepath, const std::list<std::string> &in, const std::map<std::string, std::string> &variables)
+void ImportProject::FileSettings::setIncludePaths(const std::string &basepath, const std::list<std::string> &in, std::map<std::string, std::string> &variables)
 {
     std::list<std::string> I;
     // only parse each includePath once - so remove duplicates
@@ -387,7 +405,7 @@ static void importPropertyGroup(const tinyxml2::XMLElement *node, std::map<std::
 static void loadVisualStudioProperties(const std::string &props, std::map<std::string,std::string> *variables, std::string *includePath, const std::string &additionalIncludeDirectories, std::list<ItemDefinitionGroup> &itemDefinitionGroupList)
 {
     std::string filename(props);
-    if (!simplifyPathWithVariables(filename,*variables))
+    if (!simplifyPathWithVariables(filename, *variables))
         return;
     tinyxml2::XMLDocument doc;
     if (doc.LoadFile(filename.c_str()) != tinyxml2::XML_SUCCESS)
@@ -414,7 +432,7 @@ static void loadVisualStudioProperties(const std::string &props, std::map<std::s
     }
 }
 
-void ImportProject::importVcxproj(const std::string &filename, std::map<std::string, std::string> variables, const std::string &additionalIncludeDirectories)
+void ImportProject::importVcxproj(const std::string &filename, std::map<std::string, std::string> &variables, const std::string &additionalIncludeDirectories)
 {
     variables["ProjectDir"] = Path::simplifyPath(Path::getPathFromFilename(filename));