@danmar 2017-06-26
Thanks for working on this!
As I understand it .. it might take ~10 seconds in your case to load the full solution. After it is loaded there is no overhead. So in the preprocessing/tokenizing/analysis steps the std::getenv() is not used.
Then I don't think this overhead is important.