@danmar 2017-06-26
hmm... cppcheck has a builtin unofficial functionality for showing how much time each step takes.

Example:

    ./cppcheck --showtime=summary lib

Here are the statistics when I run that on my machine on cppcheck lib:

    Tokenizer::simplifyTokens1: 183.461s (avg. 0.689704s - 266 result(s))
    Tokenizer::simplifyTokenList2: 43.6067s (avg. 0.489963s - 89 result(s))
    Tokenizer::createTokens: 17.497s (avg. 0.062267s - 281 result(s))
    Tokenizer::tokenize::setVarId: 16.6287s (avg. 0.0625138s - 266 result(s))
    Memory leaks (function variables)::runSimplifiedChecks: 9.93421s (avg. 0.11162s - 89 result(s))
    Tokenizer::tokenize::simplifyTypedef: 5.98124s (avg. 0.0224859s - 266 result(s))
    IO using format string::runSimplifiedChecks: 1.79433s (avg. 0.020161s - 89 result(s))
    STL usage::runSimplifiedChecks: 1.15679s (avg. 0.0129976s - 89 result(s))
    Uninitialized variables::runSimplifiedChecks: 0.865856s (avg. 0.00972872s - 89 result(s))
    Null pointer::runSimplifiedChecks: 0.691422s (avg. 0.00776879s - 89 result(s))
    Memory leaks (address not taken)::runSimplifiedChecks: 0.547596s (avg. 0.00615276s - 89 result(s))
    ....
    Overall time: 285.956s


Cppcheck will check all the files in lib and after that write a summary report where total time for preprocessing/tokenizer/various analyzers can be seen.

Unfortunately that option does not seem to work well when --project is used. I was not aware of this flaw.

Sometimes it take hours to analyse a solution.. and I assume that normally the "import solution" is not a bottle neck. It would be interesting with some statistics about much time "import solution" takes with and without your changes.