@danmar 2017-06-27
I applied this with 15b0eb42bdfaa6086e86be300d0b60626864e40a

Have we added you to our AUTHORS file? Otherwise.. what name do you want it to show?