@x29a 2017-06-27
Just to answer your question about timing. I put a Timer around importSln() and ran it against every configuration and against master, those are the results, so i agree that the initial overhead can be neglected (especially in release!). Thanks for merging so it will be in the next release! No need to add me to the authors, if you insist, x29a is fine.

Debug|win32
master: ImportProject::importSln: 0.021s (avg. 0.021s - 1 result(s))
branch: ImportProject::importSln: 0.023s (avg. 0.023s - 1 result(s))

Debug|x64
master: ImportProject::importSln: 0.018s (avg. 0.018s - 1 result(s))
branch: ImportProject::importSln: 0.019s (avg. 0.019s - 1 result(s))

Release|win32
master: ImportProject::importSln: 0.004s (avg. 0.004s - 1 result(s))
branch: ImportProject::importSln: 0.004s (avg. 0.004s - 1 result(s))

Release|x64
master: ImportProject::importSln: 0.004s (avg. 0.004s - 1 result(s))
branch: ImportProject::importSln: 0.004s (avg. 0.004s - 1 result(s))