@danmar 2017-06-27
well I like to have the real names. I mostly have real names: https://github.com/danmar/cppcheck/blob/master/AUTHORS.

but if you really want x29a , then that is what I will add.